feat: astarted adding nwe index
