--- pipeline/index.mjs
@@ -1,9 +1,0 @@
-import { collectData } from "./collect.mjs";
-import { convert } from "./convert.mjs";
-import { persist } from "./persist.mjs";
-
-const data = await collectData();
-
-const res = convert(data);
-
-persist(res);
