+++ pipelines/gb/configuration.mjs
@@ -1,0 +1,6 @@
+export const config = {
+  years: ["2022"],
+  propertyPrices: {
+    fetch: false,
+  },
+};
