+++ pipelines/iesg/index.mjs
@@ -1,0 +1,9 @@
+import { collectData } from "./collect/index.mjs";
+import { convert } from "./convert.mjs";
+import { persist } from "./persist.mjs";
+
+const data = await collectData();
+
+const res = convert(data);
+
+persist(res);
