+++ src/normalizers.mjs
@@ -1,0 +1,2 @@
+export const normalizeString = (input) =>
+  input.toLowerCase().replace(/\s/g, "");
